rename queries package
