+++ moonlight-core/src/main/scala/com/github/jamedge/moonlight/core/model/neo4j/queries/LineQueriesConstructor.scala
@@ -1,4 +1,4 @@
-package com.github.jamedge.moonlight.core.model.neo4j.queryconstructors
+package com.github.jamedge.moonlight.core.model.neo4j.queries
 
 import com.github.jamedge.moonlight.core.model.neo4j.Nodes
 import neotypes.DeferredQueryBuilder
